requirements.txt
pytz
tornado
aiohttp
boto3
colorama
dacite
motor
pillow
pycryptodome
pylint
rich
tqdm